ST 307 HW Assignment 2
In this assignment for ST 307, I was tasked with creating a SAS program with 17 variables with a one-way contingency table, a proc step that displays the mean, median, and mode, and a scatter plot between two variables for the first part of this SAS project. For the second part of this project, I was tasked with summarizing data from the first homework assignment graphically and numerically with useful statistics.